+++ packages/termlib/example/readline.dart
@@ -4,7 +4,11 @@ Future<void> main() async {
   final t = TermLib()..write('Type something: ');
   final input = await t.readLine();
 
-  t.writeln('${t.newLine}You typed: [$input]${t.newLine}');
+  if (input == null) {
+    t.writeln('${t.newLine}Cancelled${t.newLine}');
+  } else {
+    t.writeln('${t.newLine}You typed: [$input]${t.newLine}');
+  }
 
   return t.flushThenExit(0);
 }